requirements.txt
pyrogram>=2.0.0
pymongo==4.6.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
